create comments edition component + improvements

+++ src/components/components/postCard.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import { IoMdCreate, IoMdTrash, IoIosContact } from "react-icons/io"
 import { FaComments } from 'react-icons/fa'
+import Swal from 'sweetalert2'
 import Vote from './vote'
 class PostCard extends Component {
   onScoreUp = () => {
@@ -16,7 +17,20 @@ class PostCard extends Component {
   }
 
   onDelete = () => {
-    this.props.onDelete(this.props.post)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        this.props.onDelete(this.props.post)
+      }
+    })
   }
 
   render(){
@@ -25,14 +39,14 @@ class PostCard extends Component {
       <div className='card post shadow'>
         <div className='tools'>
           <a id="update-post" href={`/post/create-update/${post.id}`}><IoMdCreate /></a>
-          <a onClick={this.onDelete}><IoMdTrash /></a>
+          <button className="button2" onClick={this.onDelete}><IoMdTrash /></button>
         </div>
         <Vote voteScore={post.voteScore} onScoreUp={this.onScoreUp} onScoreDown={this.onScoreDown}/>
           <span className='card-body col-sm-9'>
-            <h3 className='card-title'><a href={`/post/detail/${post.id}`}>{post.title}</a></h3>
+            <h3 className='card-title'><a href={`/${post.category}/${post.id}`}>{post.title}</a></h3>
             <p className='card-text'>{post.body}</p>
             <span>
-              <p className='author'><IoIosContact /> {post.author} <br></br> {new Date(post.timestamp).toLocaleString()}</p>
+              <p className='author'><IoIosContact /> {post.author} <br/> {new Date(post.timestamp).toLocaleString()}</p>
               <p className='createdAt'></p>
               <p className='comments'><FaComments /> {post.commentCount}</p>
             </span>
